refactor: Athkar util file name
